Refactor: Reorganize folder structure and update components - Moved pages to /pages folder. - Moved Header component to /components and renamed it to Navbar. - Removed auth and setAuth props from App.js and routes.js (temporary change). - Updated Login.js, Register.js, Navbar.js, and Dashboard.js to use local state and local storage (temporary change). - Created routes.js for centralized route management.
